habitat resource improvements
This commit introduces several improvements to the hab_package
provider, and one fix for the hab_install provider.

First, set the default hab_path to "/bin/hab" in the hab_install
provider. When the Habitat "install.sh" script runs, it uses "hab pkg
binlink" to put the hab binary in the $PATH, which is by default
"/bin/hab", not "/usr/local/bin/hab". This is idempotent but not
convergent, we don't need to rerun the install every time.

Next, improve the hab_package provider. This has several components to
it.

* Ensure we `use_inline_resources`
* Validate the package name doesn't contain the version/release
* Clarify the message about the remove action
* Fix package installation convergence
* Change the test cookbook to use different package names

We're using the updated custom resource DSL in this, so we need to
`use_inline_resources` so that we get the correct behavior.

It is not idiomatic chef recipe code to specify `package` resources
with the `version` as part of the name, despite that being passed as a
string to the underlying command-line tool. While it is possible for
Habitat users to do this:

```
hab pkg install core/redis/3.2.3
```

In Chef this conflates to overloading the package name and it makes
detecting the idempotence of the resource more difficult, if not
impossible. Thus, the ability to pass the version/release to the
package name is not supported, and we validate this by raising an
exception if it's not the case.

Use the following forms for specifying Habitat package resources:

```ruby
hab_package "core/redis"

hab_package "core/redis" do
  version "3.2.3"
end

hab_package "core/redis" do
  version "3.2.3/20160920131015"
end
```

Note that when specifying only the version, and not the release, the
provider looks at the remote depot for a candidate version, and if the
installed release does not match, the package will be installed from
the remote depot. To ensure that a specific release is used, specify
it as part of the version per the last example above.

The best way to handle removing a package is to add that functionality
in the `hab` CLI, rather than do something potentially dangerous in
this provider.

Previously, the install action would reinstall packages every Chef run
because we weren't checking for the existence of the current resource.
In this commit, we have implemented `version_requirement_satisfied?`
to determine whether there is a current version installed and if it
matches the version specified in the new resource. Because the Habitat
packages have a "version" and a "release" that both combine to be a
valid version identifier, it can be tricky to determine reliably which
version is installed, so we do our best here by splitting the
identifier on `/`.

Because we no longer accept versions or version/releases as part of
the package name, we need to use unique package names in order to
avoid resource cloning issues.

Signed-off-by: Joshua Timberman <joshua@chef.io>
Signed-off-by: jtimberman <joshua@chef.io>