Add property for ExecStart options
In some cases, it is desirable to change only the ExecStart parameter
from the default value. Before this commit, one would have to write all
the unit content out. With this, options can be passed in using an array
or a string. If the property is a string, it gets Array()'d.

Signed-off-by: Joshua Timberman <joshua@chef.io>