Fix linting errors
We used chefstyle because the intent is that these resources and
providers would be ported to core Chef at some point. However, given
the rapid development of the upstream Habitat project, we won't be
able to put them into Chef yet, so we'll keep them in the cookbook for
now until the upstream APIs and tools are under less churn.

It isn't difficult to change to chefstyle when we put them into Chef,
though, so let's cookstyle this up.

Signed-off-by: Joshua Timberman <joshua@chef.io>