Add a hab_config resource
Add hab_config tests to travis
Add config test suite to travis kitchen config
Add a hab_config resource
add toml gem, require later chef

Ideally we can support some chef 12, but we do need a semi-recent one
for changes in methods that are defined in the classes that we're
using. It turns out 12.20.3 is the lowest version that we can use
without rewriting several parts of this cookbook, so here we are.

We also need to install the toml gem to ensure it is present on these
older versions of chef.

Signed-off-by: Mark Harrison <mark@mivok.net>
Signed-off-by: Joshua Timberman <joshua@chef.io>