Pin the version of hab we install
We need to pin the version of habitat installed with the `hab_install`
resource. This cookbook is developed against the specific version of
Habitat mentioned, and makes reasonable attempts to work with older
versions. However as Habitat itself is a pre-1.0 project at this time,
APIs (both its publc API and the CLI API) may change at any time and we
need to account for potential breaking behavior.

The cookbook version is changed as a suggestion that we more closely
track this cookbook's version with the upstream Habitat version. The
README indicates the places where the version needs to be updated.

Signed-off-by: Joshua Timberman <joshua@chef.io>