create user in hab_install
Fixes #54

This change adds a property to the `hab_install` resource that will
automatically create a `hab` system user. This is because many
services in Habitat's core plans run as the "hab" user, and we want to
avoid services failing to start or failed chef client runs because the
user wasn't there.

End users can opt out of the user creation by setting the property to
false, and create their own user in their recipes if their service
runs as something that isn't `hab`.

Finally, all the double quotes in the README.md are single quotes, so
our examples pass cookstyle ;)

Signed-off-by: Joshua Timberman <joshua@chef.io>