Ignore failure when installing from acceptance
The Habitat bldr for acceptance isn't monitored and can sometimes be
down when we want to run tests in Travis. This shouldn't stop us from
doing our tests, so ignore failures. Also remove the test that is done
post convergence, as we may not have that package installed, and we
can check other packages are installed from that test as using an
alternative URL is not required to know that the functionality is
working.

Signed-off-by: Joshua Timberman <joshua@chef.io>