Add: bldr_url property for hab_sup
Because the `hab_sup` resource explicitly ensures that the supervisor is
installed, it includes a `hab_package` resource.

However, for users in a private environment, with an on-prem Builder, it
is possible that access to upstream/public Builder is neither possible
nor desirable.

This adds the ability to specify the `:bldr_url` property to `hab_sup`,
which is then passed through to the `hab_package` resource.

Signed-off-by: Stephen Nelson-Smith stephen@atalanta-systems.com